KStarsData should not call methods on SkyMap, but only emit signals.
KStarsData is a class that handles the simulation state and data. We
would not want it to have anything to do with SkyMap, which is a way
of rendering the view of the sky, and allowing the user to interact
with it.

So instead of calling methods on SkyMap and having a SkyMap pointer
passed around to its methods, KStarsData should only emit signals
every time the sky map needs to be updated.

This is just one baby step towards the ultimate, grand goal of
decoupling the computational aspects of KStars from the interface,
that many of us have had in mind. I might consider attempting pieces
of this gargantuan task since I have a personal interest -- the dream
of running a KStars engine on a server without X11, with an interface
that is web based, so that KStars can run as an astronomy computation
service. Whether this dream will be a reality or not is to be seen,
and its success is severely constrained by what little time we
developers have at our disposal.

What I want to start with, is to build on the old "--dump" command
line option that we had working in the past (doesn't seem to be
working now), that would dump an image of the sky without opening the
KStars window etc. However, the old way of doing dump required
SkyMap. Really, SkyMap should not be in the picture. If all I want is
an image of the sky, the user is not in a position to interact, so why
have an interactive SkyMap involved?

Comments and contributions towards this goal are welcome. I hope this
commit is one of many that will lead towards this goal.

CCMAIL: kstars-devel@kde.org