Adding new Pluto photograph just taken recently by NASA New Horizons spacecraft!
Also making a couple of modifications to the asteroid class to enable this, but we would probably need a better solution so that we can include images for other asteroids as well. Will work on that. The Pluto image is scaled now exponentially to allow the user to see
the full image at a reasonable resolution on full zoom.

CCMAIL:kstars-devel@kde.org