Improve common names for NGC/IC objects
Improved common names for many NGC/IC objects.

Removed some incorrect common names / ambiguous common names that
could be confused with other objects.

Added many new common names.

Added names for fragments of the Veil nebula, some groups and chains
of galaxies etc.

One common name 'Electric Guitar Cluster', is not to be found
elsewhere, and is an introduction of mine after a friend's calling it
that. It does indeed possess a semblence to an electric guitar, and it
is not uncommon for amateur astronomers to introduce colloquial names
for objects (this is the norm rather than the exception). It is much
like our previous introduction of "Galaxy Not Found" for NGC 404. If
you disagree, please feel free to revert this.

Note: This does not affect the Steinicke catalog. Dr. Steinicke does
not permit us to alter his catalog.

CCMAIL: kstars-devel@kde.org