Use red colored link text in help labels when using night color scheme
A lot of astronomers would use a red acrylic screen filter when using
KStars on the field / in the observatory. The red acrylic screen
blocks blue and green light, whereby any text that is blue and green
appear black through the screen. Thus, any text in blue/green over a
black background is invisible.

This was the case with the KSHelpLabel links in (for example) the
details dialog, they were impossible to see through a red-screen
filter.

This commit fixes this problem by choosing red colored links when in
the night vision color scheme.