Use alpha-blended gradients in Alt. vs. Time
Use alpha-blended gradients in the Alt. vs. Time so that we can
include both solar and lunar interference.

It seems to me that the gradients should start at sunset and end at
the end of astronomical twilight to be true to the nature of the
darkening. In the first place, using a linear gradient for a
logarithmic decrease in brightness is questionable, but may be
justifiable by the logarithmic response of the eye. So of course, the
entire representation is not scientifically accurate, but it seems
more satisfying to me to let the exact times determine the start and
stop of the gradients rather than having a 20 px extra transition
region like earlier. The removal of the +20 -20 values is to achieve
this.