Do not specify false time-zone information in the popup dialog
Earlier, the long format for time was used, which causes the time
representation to include the timezone, which inevitably happens to be
the timezone associated with the locale. However, KStars is designed
to work at any geographic location, and when it does so, it reports
time in local time. Thus, if using KStars at a geolocation that is not
the local time, that time-zone spec is simply invalid.

BUG: