Questionable(!) fix in KSAlmanac to get moon times correct for the night
An uncertain / questionable fix in KSAlmanac that seems to get the
correct set of moon times that an amateur astronomer would be
interested in.

But some times, this has been shown to produce whacky / buggy
results. Needs further investigation to fix.

Also, what solar times this produces, is an entirely different
question to be answered.

As far as I know, this doesn't fix any bugs, but likely does not
introduce any more than are already present.