Show the current time in the correct timezone in alt. vs. time plots
The line that shows the current time in the alt. vs. time widget
requires geo location info since the chosen location for the
alt. vs. time computation need not be in the same timezone as the
computer's configuration.

In fact, the X-axis of the alt. vs. time plots is Local Time, and it
is meaningless to show the current time in an arbitrary timezone as
the local time.

This inclusion of geolocation information as a pre-requisite to draw
the line ensures that we never provide false information on the
current time and timezone.

Also, another fix is that the false timezone information in the
labeling of the mouse-click position is removed.

BUG: