Endow Alt. vs. Time with the ability to plot sky glow due to moon
This is a backward-compatible change, since if no moon information is
supplied, the moon gradient is simply not displayed.

Otherwise, if the moon info is supplied, the Alt. vs. Time widget
shows a gray gradient indicating the duration and intensity of skyglow
from the moon.

This commit is the implementation in the widget. A further commit will
use it in the observing list.