Use smooth scaling in eyepiece view feature. Also show temp images.
1. Some cosmetics
2. Using smooth scaling in the eyepiece view feature to produce a less
   granular image.
3. Show image in Eyepiece View feature even if it is just a temporary
   image

Some TODO items and future thoughts:

This image handling seems broken in the sense that the concept of a
temp image is confounding to a developer; then imagine how confounding
it must be to a user. It was a suggestion put forth by yours truly to
ensure reduced consumption of disk space, but it seems superfluous
because the "temp" images seem to be retained anyway.

A rewrite of the image handling, to ensure that the images are always
saved, with a feature to "Download all images" instead of "Save all
images", and also including metadata of the image somehow or the other
(things like DSS vs. SDSS, field of view size w x h, and which UBVRI
band etc.) and a way to at least check by back-tracking if a point is
in the latest SDSS footprint, also moving to the latest SDSS data
release would be very nice.