Make the eyepiece field grab render smoothly despite rotation etc.
The grab of the skymap is now rendered on an SVG file and then the SVG
is rendered on a QImage that is 4 times bigger than what is actually
needed. This allows for very smooth rotation and scaling with Qt's
smooth transformation algorithms.

Copying Jasem because he uses this in Ekos.

CCMAIL: mutlaqja@ikarustech.com