Disentangle dependencies to ensure faster compilation
Almost all of KStars would recompile everytime a change was made to
ObservingList, because of the inclusion of observinglist.h in
kstars.h! Removing that revealed a long chain of unmet includes and
missing forward declarations, which are fixed. Also required moving a
bunch of methods that shouldn't have been expanded inline into the
.cpp file.