Make sure that the DSS image is rotated to match even when using Alt-Az
In Alt-Az mode, altitude is vertically up. DSS images always come with
north up. If we get a snapshot of the sky map in Alt-Az mode, the DSS
image needs to be rotated to match the sky map. To do this, we use
spherical trigonometry to calculate the angle through which we must
rotate and then rotate.

Needs a lot of testing before it can be confirmed to work.