Include preset settings in Eyepiece View for Dobsonians and Refractors
Including two preset settings in Eyepiece View that attempt to
reproduce the exact orientation of the field as seen through
refractors without erecting prisms, and through Dobsonian telescopes
that have a focuser on the side. These presets are meaningful only
when KStars is in Alt-Az mode (using Horizontal Coordinates) as of
now, but should be easy to get working in both modes by including the
equatorial-to-horizontal contribution to rotation.

The Dobsonian setting has been field-tested using an Obsession
telescopes 18-inch which has a focuser on the side. The refractor
setting has not been field-tested.

The Dobsonian setting will not work out-of-the-box for Dobsonian
telescopes that have a focuser at an angle (this is usually the case
with <= 12 inch Dobsonians).