Remove the notion of temp images in Obs. Planner. All images are saved.
The concept of temporary images that are deleted upon exit seems both
counter-intuitive and superfluous. This commit gets rid of the concept
of temp images. Images that are downloaded in the observing list are
now saved on disk, and may be deleted either individually by clicking
the "Delete" button, or collectively by clicking the "Delete all
Images" button.

This change is NOT backwards-compatible. Any DSS images that were
saved as Temp_Image_<Obj. Name> are no longer visible / available in
KStars. In fact, they must be manually deleted from the disk by
opening the appropriate directory.

This backwards-incompatibility is partly intentional, since we
introduced the idea of DSS metadata. In any case, it would be wise to
re-download the DSS images with proper metadata so that features like
eyepiece view, and in future, DSS overlay work correctly without
trouble.

CCMAIL: kstars-devel@kde.org
GUI: GUI elements / config. options have been removed.