Add DBus interface method to set time to the current time.
Adding DBus interface method org.kde.kstars.setTimeToNow which sets
the time to the current time. This allows for syncing of the
simulation clock with the system clock.

Example use-case:

Running the simulation clock continuously takes a lot of CPU, whereas
for many practical purposes, it might suffice to sync the clock every
minute or so automatically. In these cases, one could simply run a
script on a terminal to achieve this:

while true; do
   qdbus org.kde.kstars /KStars org.kde.kstars.setTimeToNow
   sleep 60;
done;

This will ensure that alt-az coordinates are reasonably up-to-date,
eyepiece view is reasonably up-to-date etc. without taking up a lot of
CPU.

CCMAIL: kstars-devel@kde.org