EyepieceField: Allow downloading DSS when no image has been supplied
This is particularly useful when launching Eyepiece View from the sky
map popup menu, since no DSS image is then supplied to Eyepiece
View. A "Fetch DSS" button is displayed which may be used to download
an image from DSS and then render it in Eyepiece View. The file is a
temporary file, that is reused and finally removed when EyepieceField
is destroyed -- so the DSS image is not saved.

Future work would include the option to save downloaded DSS images to
make a library, to avoid re-downloading. Eventually, this library
could be used for DSS overlay.

FEATURE: Allow fetching of image in eyepiece view when none is supplied
GUI: A new button to fetch DSS images has been added to Eyepiece View.