Move the code for rendering the eyepiece view into a static method
The code that takes care of orienting the images (rotation, flip,
invert etc.), inverting colors on the DSS image, and overlaying the
image on the chart has been moved to a separate static method
renderEyepieceView.

EyepieceView::render() essentially calls this method with the state of
the checkboxes to render the view, and then displays it.

This allows us to use these features outside of the Eyepiece View
tool, for example in Ekos, or from DBus.

DBus bindings will allow for more powerful star-chart generation etc.

CCMAIL: mutlaqja@ikarustech.com