Add dialog to export eyepiece view, w/ annotations to help orient chart
This commit adds an "Export" button to Eyepiece View which opens a
dialog where one may add a bunch of markings helping the user to
orient the produced eyepiece view for various times of observation.

This feature has not been field tested.

Example output:
http://wstaw.org/m/2016/01/18/M81.png

The markings indicate how to orient the chart for various (local)
times in Austin, TX when viewing through a Dobsonian with a focuser
parallel to the ground. If observing at 3:50 AM, the chart is used as
it is, but for example, at 2:20 AM, the chart needs to be rotated by
about 45 degrees so that the tick corresponding to 2:20 AM points
upwards.

CCMAIL: kstars-devel@kde.org