Fix precession bug: Reset lastPrecessJD when star data is reinitialized
We now recompute precession corrections only if there has been a large
shift in time.

With dynamically loaded stars (DeepStarComponent), the star object is
recycled in memory to avoid the overhead of memory allocation and
construction. The new star data is re-initialized in a pre-existing
StarObject by calling StarObject::init() with the new data.

Since we added the lastPrecessJD much later, it has to be initialized
in StarObject::init(); otherwise, the catalog coordinates will not be
precessed when a previously existing StarObject is re-purposed for a
new star.

CCMAIL: kstars-devel@kde.org, mutlaqja@ikarustech.com