Fix some focus bugs by updating focus coordinates appropriately
It has been noted earlier that KStars does not center the map
correctly on first shot in many cases. The offset from the actual
object and the map center was consistent with the possibility that it
could be a result of not handling precession correctly.

This patch is likely to fix this bug although it needs more
testing. In any case, one must call SkyPoint::updateCoords()
appropriately before, so this patch probably does no harm. In a
cursory test, KStars centered correctly on the location of the object
without any offset.

Note: this seems to produce some warnings, so one should double check
that the updateCoords from subclasses of SkyPoint are being called
when needed.

CCMAIL: mutlaqja@ikarustech.com