A SkyComponent which loads a custom catalog, but can also add objects.
Adding SyncedCatalogComponent, a new SkyComponent that loads a custom
catalog, and can also add objects synced with the database.

Custom catalogs are stored in an SQLite database. CatalogComponent
loads these into memory and handles their rendering etc. This new
subclass also allows adding of new objects at run time. The new
objects are written to the database and also available at run-time,
and hence the "synced".

Typically, one would not add objects to a well-known catalog, so
SyncedCatalogComponent will usually be used with "fake" catalogs, just
lists of objects that we put into a catalog with some name and assign
IDs to the objects, that are completely unofficial. Since the catalogs
are 'unofficial', the Catalog_Prefix + Object ID name doesn't mean
anything, so we try to hide it.