Add a new button to the Find Dialog, that resolves using internet
Without the button: resolution was possible only when the search
filtering did not return any results. Suppose you have "Shk 166"
already in the database, but don't have "Shk 16". Typing just "Shk 16"
will show and select "Shk 166" in the filtered list, with no way of
asking KStars to query Sesame for "Shk 16".

To allow for this, this commit adds a new button that enables internet
search when the list has results, but an exact match has not been
found. If an exact match exists in KStars, the button is grayed out to
prevent multiple copies of the same object from forming. If no search
text is available, once again the button is grayed out. The default
action of hitting Enter is still as before.

The button only appears if the resolution option is selected in the
configuration dialog. The checkbox in the config dialog is still
useful as this can be "messy" for some users who don't want to pollute
their sky map with a bunch of wrong search results.

CCMAIL: afedoskin3@gmail.com, kstars-devel@kde.org
GUI: