Only add "kstars" directory if using QStandardPaths::GenericDataLocation
In my current opinion,

ideally, we should not be doing this. We should instead use
QStandardPaths::AppDataLocation within our code, and reserve
GenericDataLocation only when we actually need to store something that
is shared across applications. That permits us to actually use
QStandardPaths::GenericDataLocation when we actually need it.

This patch fixes misbehavior with respect to image handling in the
Observation Planner.

CCMAIL: afedoskin3@gmail.com, kstars-devel@kde.org