Code to put USNO NOMAD binary data file contents into a SQLite database
Note: This is an initial version that does not use transactions or the
like, and is VERY VERY SLOW. It'll take A YEAR on my computer to
process the full file! I'm obviously going to make improvements soon.

N.B. Following StackExchange post might be useful:
http://stackoverflow.com/questions/1711631/improve-insert-per-second-performance-of-sqlite

CCMAIL: mutlaqja@ikarustech.com