Make KSNumbers::obliquity() return CachingDms
Profiling did not reveal a significant advantage here mostly because
of the updateCoords() short-circuit in
StarObject::JITUpdate(). However, I presume that the impact will be
significant when slewing / zooming -- operations that are very hard to
profile, and time steps of > 1 minute, which we will have to
unfortunately address only later.