Fix the definition of Julian Epoch in KStarsDateTime
Since epochs are a measure of uniform flow of time, they cannot be
affected by leap years etc.

The peer-reviewed definition of Julian Epoch that this patch uses
comes from Eric Weisstein's World of Science:
http://scienceworld.wolfram.com/astronomy/JulianEpoch.html

(This definition may also be found on Wikipedia). The constant
2451545.0 is identical with the #define J2000 -- i.e. the Julian date
corresponding to the J2000.0 epoch.