Update comments in SkyPoint::precess()
Rotation matrices are most efficient, even though access time is long

See:
https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation#Performance_comparisons

Applying a rotation matrix, which costs 9 multiplications, is more
efficient than applying a quaternion on a vector. The only issue is
the overhead of accessing 9 double precision elements.

It looks like it is best to stick to rotation matrices.