Restruct. Obs. Planner: More vertical real estate for the Table Views.
It seems more natural to provide more vertical real estate for the
Table View objects, since the list spans a fairly long length in
typical cases. Just showing 4 rows is not very useful, but showing as
few as 2 columns should still be useful (it takes less scrolling to
scroll through the small number of columns).

However, we would want to make the first column, the one with the
names of objects, always visible irrespective of how we scroll
horizontally. To this end, there is a solution in the form of a Qt
example which we can use in KStars with minimal modification:

http://doc.qt.io/qt-5/qtwidgets-itemviews-frozencolumn-example.html

N.B. The BSD License used in the example is GPL-compatible.

Other improvements necessary are:

1. Put the list on the left -- feels more natural.

2. Move the time schedule option in the Session View below the table
   view.

3. Gray out the date box in WishList mode to indicate its irrelevance?
   It is technically not irrelevant since the RA/Dec shown are
   apparent coords! Maybe this should change.

4. Show some vital info about the object in a label

5. Make sure buttons don't expand if we resize -- only the table view
   must expand to fill up as much real estate as possible.

6. Allow the user to hide / show columns in the TableView.

7. Make sure the "Minify" option works well.

8. Would be nice if we could "minimize" the observing list to an icon
   in the bottom-right of the main window.