Show the correct epoch in Details Dialog, and also the correct coords.
Fixes a bug where the (RA, Dec) for simulation clock epoch was
identical to the J2000 coordinates in the details dialog, if the
object was not rendered on the screen. This is relevant, for example,
when invoking the details dialog from places other than the sky map
such as the Observation Planner.

The corrections are as follows:

1. Call updateCoords() on the Sky Object with the current update
   KSNumbers.

2. Set the epoch from the coordinates that we are actually displaying,
   i.e. using lastPrecessJD.

CCMAIL: mutlaqja@ikarustech.com

BUG: