Avoid truncating the wishlist if there's a crash while saving it.
There is a bug which induces a crash when writing the observing wish
list, and ends up truncating the wish list. This can be very
frustrating because of the loss of the wishlist.

Although the bug still needs to be found and fixed, this change in
code at least prevents the wishlist from being truncated upon crash.