Fix regressions in rendering of Eyepiece View in Observation Planner
These were presumably caused because of the use of clones of the
original SkyObjects in ObservingList (see
f11601a3f2a7dd99f1fba49425372baf8cc462fc). These cloned SkyObjects
don't have updateCoords or EquatorialToHorizontal called on them.

This was causing two problems:

1. The Sky map "screenshot" would not center correctly on the object,
   but would instead always be off by a few arcminutes.

2. The DSS image would have an indeterminate rotation angle and
   thereby would not be rendered at all.

This patch fixes these two issues, but has potential for introducing
regressions in the Eyepiece View as invoked from the sky map
(KSPopupMenu) or in the export option with ticks.