Fix spurious timezones in time infobox display
Before this patch, one would see things like:

UT: 12:37:32 PM CDT

Clearly, this is wrong. Similarly, there is no reason that the
timezone that KStars is using (based off of geolocation) matches that
of the system locale. So both LT and UT would carry a spurious
timezone name (i.e. that of the system locale).

This commit:

1. Standardizes UT to HH:mm:ss -- it would be ugly to use AM/PM etc
   for UTC.

2. Removes timezone from LT display. There seems to be no clean way to
   do this in an otherwise locale-respecting manner, so we get the
   format string using QLocale::timeFormat() and then remove the
   timezone format symbol 't' from the string using QString::remove,
   then pass it to QTime::toString.

BUG: