Since several people have had problems with the default directories for indiserver, astrometry, xplanet, etc on OS X when they build KStars themselves, I thought I would write some code to fix the problem. Note that the problem has only been experienced by people who build kstars themselves and either don’t have all the internal content in their app or they delete their kstarsrc file without deleting their data directory. In both of these situations they end up with a problem because the defaults in options are for Linux and are not correct on OS X.  If the user doesn’t run the code that copies the data directory or the user doesn’t toggle the internal checkboxes in the options, then the Linux defaults are written to the kstarsrc file instead of the defaults for OS X.
So I worked this evening to try to come up with a way that the default paths that do not rely on the internal structure of the app, but instead use the default locations of the external packages on OS X could be written to the user’s kstarsrc file by default.
Note that this change will not affect the majority of users who will probably be installing kstars from the dmg and won’t have a data directory yet, so they will still get everything set to internal which will work just fine for them.
This change should mainly just affect people building kstars themselves from source.  Of course, for those people, they will still need to install all of these items to their proper locations.

CCMAIL:kstars-devel@kde.org