Junion Job: Investigate NEOs is fixed!
I've constructed a request to JPL myself and got out a csv of several hundred megabytes, which is certainly not suitable for KStars.
Looking at the differences between my query and the one in KStars, I found some extra parameters, which had not been included in my one: c1_item=Ai and c1_op=%3C and c1_value=12.
That basicly means, that JPL filters the asteroids for the ones which have an parameter 'H' (absolute magnitude) smaller than 12, which means that they are 'visible'.

Added "Waiting for Server Response" to filedownloder class to improve usability.

CCMAIL:kstars-devel@kde.org