I've added some simple comet tails and an option to disable them.
There is a bit of Fuzz, cause astyle was doing it's thing. (Has anyone
run astyle recursively yet? There are still some unformated files.)

To clarify things: I just added a "drawComet" method to the skypainter,
an option to the kcfg and the UI and one tweak to retrieve an instance
of the sun.

To note: the findPA method could be tweaked, by somehow caching the Suns
position, because by now it is re-projected for every comet tail. (Well
that's actually OK, because you hardly see more then two at a time.)