For quite some time, I have known there was an issue with the SimClock in KStars.
Whenever you try to run it in hours or anything longer than that, the interface becomes much more unresponsive.  I figured it was an issue with an animation thread not sleeping long enough for the interface to have a chance to do other things.  I investigated that today and found that I was close.  There is a QTimer with a time interval that is used for the time simulation normally, but if you go beyond a certain speed, it just set the clock and didn't use a timer.  It seems to be called "manual tick."  Well the problem is that it needs to wait a moment before doing that so the interface can respond to any other commands.  I found the solution was very simple, a one shot timer, even one as short as 1 millisecond seems to solve the problem.

CCMAIL:kstars-devel@kde.org