Migrating KStars to use QLoggingCategory framework which supports granular control over logging in runtime.
The migration is still not complete yet but the base structure is there and operational and we should migrate any qDebug..et al when appropiate to use the QLoggingCategory framework. When you write code, you must abide by this, I will not accept any code that does not use the framework when logging.

BUGS:383177
CCMAIL:kstars-devel@kde.org