This patch should fix two bugs I found with PHD2.
The first is that it would state “Calibration Completed” when it first connects even when PHD2 had not even started calibrating yet.  Basically when first told to guide, all guiders would  be told to calibrate and PHD2’s method would just emit that the calibration was successful then move on to guiding when the calibration would actually occur.  To solve this, I just prevent it from trying to calibrate PHD2 in the first place and just go straight to guiding, where PHD2 will automatically calibrate it.  Now it doesn’t print that the calibration is completed until it really is completed.

The second is an issue with the situation where PHD2 is already guiding when KStars connects to it.  Multiple state commands are issued, that it is guiding, that it is connected, etc.  The issue was that the Guide button was being disabled because it was guiding, then enabled because the equipment was connected, and then it tried to disable it again because it was guiding but it couldn’t because the first time it did it the PI was set to animated.  So I disabled the lines of code that prevented the setBusy command from executing if the PI indicator is running, but only for PHD2.  Now the Guide button works correctly.

I also deleted an unnecessary line of code from the patch I made yesterday that added the pulse graphs since the line above it already takes care of it.