After a bunch of work today and several failed approaches, I finally seem to have fixed the bugs with the changes in theme on Mac.   The issue was that a number of objects including some QLineEdits, QSpinBoxes, QDoubleSpinBoxes, dmsboxes, QComboBoxes, TabBars, and a few QPushButtons were not being updated correctly when the theme changed.  Some were being changed to the current theme when you changed to the next theme, some were keeping the same theme they loaded with the whole time and other problems.  I did have a couple of solutions that worked but were overly complex.  The big problem was that they seemed to not want to accept the same QPalette that I passed to the application no matter what I did, but they were very happy accepting another QPalette.  Then I stumbled across this:
https://stackoverflow.com/questions/47971350/change-application-palette-run-time-without-restarting-the-application

QPixmapCache::clear();   And that solved most of the issues.  There were still a couple of issues.  I resolved those by reordering some of the commands first, and for dmsbox and QTabBar, I set their palettes separately because the change did not propagate down to them properly.  (This was originally my solution for all of the issues, but most of them were solved with the link above, which works better.)

So now, the issues with switching themes should be solved.  It was a bigger issue than I expected it to be.

Please test on Linux, and feel free to Ifdef anything that needs to be Mac only.